// packages/ui-components/src/styles/styles-config.d.ts
export declare const stylesConfig: {
    color: {
        white: string;
        guardsmanRed: string;
        fiord: string;
        tango: string;
        wildSand: string;
        textColor: string;
        doveGrey: string;
        azureRadiance: string;
        black: string;
        mountainMeadow: string;
        primary: string;
        secondary: string;
        error: string;
        warning: string;
        success: string;
    };
    spacing: {
        extraSmall: string;
        small: string;
        normal: string;
        large: string;
        big: string;
        huge: string;
    };
    fontSize: {
        extraSmall: string;
        small: string;
        normal: string;
        large: string;
    };
    opacity: {
        disabled: number;
    };
    fontFamily: {
        primary: string;
    };
    fontWeight: {
        bold: number;
        bolder: number;
    };
    boxShadow: string;
    zIndex: {
        sidebar: number;
        dropdowns: number;
        overlays: number;
    };
};